I have to complete some minor refactoring in the the functions that prepare the data for visualization.
Somehow I had forgotten that I had created the selectedEntries array to only hold the ids of all the selected weather data. When I finally got to the chart creation part I had assumed that I set it up to store all the weather data and not just the ids.

Now I have to just change a little bit of the nested orEach statements but i'll do it in separate commits to better keep track of the changes.