Earlier I had separated the concerns for border updation by moving the weather entries to a separate file once again, but without the added complexity of animating the entering and exiting  of the entries from the viewport.
I thought that by doing so, memoization won't prevent the re-rendering of the selected entry but it was still doing so.

This time I'm trying it without memoization to see if the scroll position gets drastically affected when the user selects an entry. The border colour will obviously update in this manner but I'm skeptical about the scroll position being maintained on account of the entire list component re-rendering this time